\name{Nurses}
\docType{data}
\alias{Nurses}
\title{
Nurses' stress in different hospitals
}
\description{
This dataset comprises three-level simulated data extracted for a hypothetical study investigating stress levels within hospital settings.
The dataset focuses on nurses working in specific wards within various hospitals.
It includes several variables, such as nurse age (measured in years), nurse experience (measured in years), nurse gender (\code{0} for \code{male}, \code{1} for \code{female}), ward type (\code{0} for \code{general care}, \code{1} for \code{special care}), and hospital size (\code{0} for \code{small}, \code{1} for \code{medium}, \code{2} for \code{large}).
The dataset in the package is obtained from the original dataset by leaving out some of the unused columns.
}
\usage{data(Nurses)}

\source{{https://multilevel-analysis.sites.uu.nl/datasets/}}

\references{
Hox, J., Moerbeek, M., and van de Schoot, R. (2010). \emph{Multilevel Analysis: Techniques and Applications}, Second Edition (2nd ed.). \emph{Routledge}. <https://doi.org/10.4324/9780203852279>
}
\keyword{datasets}


