\name{BFI-package}
\alias{BFI-package}
\docType{package}

\title{Bayesian Federated Inference}

\description{
Bayesian Federated Inference method combines data from different (medical) centers without sharing them. In this version of the package, the user can fit models specifying Gaussian and Binomial (Logistic) families.
}
\details{
\tabular{ll}{
Package: \tab BFI\cr
Type: \tab Package\cr
Version: \tab 0.5.2\cr
Date: \tab 2023-07-16\cr
License: \tab GPL (>=2)
}
}
\author{
Hassan Pazira\cr
Maintainer: Hassan Pazira \email{hassan.pazira@radboudumc.nl}
}
\references{
Jonker M.A., Pazira H. and Coolen A.C.C. (2023). \emph{Bayesian Federated Inference for Statistical Models}, \emph{Statistics in Medicine}, Vol. 0(0), 0-0. <https://doi.org/10.48550/arXiv.2302.07677>
}

\keyword{Bayesian}
\keyword{Federated}
\keyword{package}
