\name{inv.prior.cov}
\alias{inv.prior.cov}
\title{Creates an inverse covariance matrix for a Gaussian prior}
\description{
\code{inv.prior.cov} builds a matrix which can be used as the inverse of a covariance matrix for a Gaussian prior distribution, to be used in the main functions \code{MAP.estimation()} and \code{bfi()}.
}
\usage{
inv.prior.cov(X, lambda = 1, family = gaussian,
              intercept = TRUE, independ = TRUE,
              set_seed = NULL)
}
\arguments{
\item{X}{design matrix of dimension \eqn{n \times p}, where \eqn{p} is the number of covariates (predictors).}

\item{lambda}{the tuning parameter for ridge (L2) penalization that controls the strength of the ridge penalty. It could be either a single positive number, a vector of two elements, or a vector with length equal to \eqn{p}, \eqn{p+1} or \eqn{p+2} (where \eqn{p} is the number of covariates, i.e., number of columns of \code{X} argument) depends on the arguments \code{X}, \code{family}, and \code{intercept}. If \code{lambda} chosen by the user is a single positive number, the function \code{inv.prior.cov()} consider \code{lambda} as a vector whose all elements are equal to that scalar value. Default is \code{lambda=1}. If the user choose a vector of only two elements as an entry value, the \code{inv.prior.cov()} function set \code{lambda} as a vector whose all elements, except the last one, are equal to the first element of the entry, and the last element is set to the second element of the entry. Used when \code{independ=TRUE}.}

\item{family}{a description of the error distribution and link function used to specify the model. This can be a character string naming a family function or the result of a call to a family function (see \code{\link{family}} for details). By default the \code{gaussian} family (with identity link function) is used.}

\item{intercept}{logical flag for fitting an intercept. The intercept is fitted if \code{intercept=TRUE} (the default) or set to zero if \code{intercept=FALSE}. If \code{intercept=FALSE}, the first row and column of this matrix are set to zero}

\item{independ}{logical flag for creating the prior covariance matrix with (in)dependent variables/parameters. If \code{TRUE} (the default), the result is a diagonal matrix which means the parameters are independent. If \code{FALSE}, the off-diagonal elements of the resulted matrix are generated from the standard normal distribution \code{rnorm()}, and the diagonal elements are the vector \code{lambda}.}

\item{set_seed}{if \code{independ=FALSE}, this argument (which should be an integer) ensures consistent results. If \code{set_seed=NULL} (the default), the results are not consistent.}
}

\details{
\code{inv.prior.cov} creats a matrix whose dimension depends on the arguments \code{X}, \code{family}, and \code{intercept}. If we assume the parameters are independent (\code{independ=TRUE}), the function \code{inv.prior.cov()} returns a diagonal matrix with the vector \code{lambda} as its diagonal. If \code{independ=FALSE}, \code{inv.prior.cov} returns a matrix which is equal to \code{z\%*\%t(z)} where z is a matrix such that all elements are generated by standard normal distribution.
}

\value{
\code{inv.prior.cov} returns a matrix. The dimension of the matrix depends on the arguments \code{X}, \code{family}, and \code{intercept}.
}

\references{
Jonker M.A., Pazira H. and Coolen A.C.C. (2023). \emph{Bayesian Federated Inference for Statistical Models},
\emph{Statistics in Medicine}, Vol. 0(0), 0-0. <https://doi.org/10.48550/arXiv.2302.07677>
}

\author{Hassan Pazira\cr
Maintainer: Hassan Pazira \email{hassan.pazira@radboudumc.nl}
}

\seealso{
\code{\link{MAP.estimation}}.
}

\examples{
#-------------
# y ~ Binomial
#-------------
X       <- data.frame(matrix(rnorm(50 * 4), 50, 4))
lambda  <- 0.05
# We assume the same (inverse) covariance matrix for all
# locations and consider independency of prior parameters
(Lambda <- inv.prior.cov(X, lambda, family=binomial))
# No intercept
(Lambda <- inv.prior.cov(X, lambda, family="binomial", intercept = FALSE))

#-------------
# y ~ Gaussian
#-------------
X       <- data.frame(matrix(rnorm(50 * 3), 50, 3))
lambda  <- 0.01
sigma2e <- 0.5
# If we consider dependency for all prior parameters:
(Lambda <- inv.prior.cov(X, family="gaussian", independ = FALSE, set_seed=1123))
# No intercept
(Lambda <- inv.prior.cov(X, family="gaussian", intercept = FALSE, independ = FALSE, set_seed=1123))
}
\keyword{Bayesian}
