\name{Nurses}
\docType{data}
\alias{Nurses}
\title{
Nurses' stress in different hospitals
}
\description{
This data set contains three-level simulated data from a hypothetical study on stress in hospitals. The data are from nurses working in wards nested within hospitals. It is a cluster-randomized experiment. In each of 25 hospitals, four wards are selected and randomly assigned to an experimental and a control condition. In the experimental condition, a training program is offered to all nurses to cope with job-related stress. After the program is completed, a sample of about 10 nurses from each ward is given a test that measures job-related stress. Additional variables are: nurse age (years), nurse experience (years), nurse gender (\code{0 = male}, \code{1 = female}), type of ward (\code{0 = general care}, \code{1 = special care}), and hospital size (\code{0 = small}, \code{1 = medium}, \code{2 = large}).
}
\usage{data(Nurses)}
\source{{https://multilevel-analysis.sites.uu.nl/datasets/}}
\references{
Hox, J., Moerbeek, M., and van de Schoot, R. (2010). \emph{Multilevel Analysis: Techniques and Applications}, Second Edition (2nd ed.). \emph{Routledge}. <https://doi.org/10.4324/9780203852279>
}
\keyword{datasets}


