\name{trauma}
\docType{data}
\alias{trauma}
\title{
Trauma patients from different hospitals
}
\description{
This data set consists of data of 371 trauma patients from three hospitals.
The binary variable \code{mortality} is used as an outcome, and variables \code{age}, \code{sex}, the Injury Severity Score (\code{ISS}, ranging from 1 (low) to 75 (high)) and the Glasgow Coma Scale (\code{GCS}, which expresses the level of consciousness, ranging from 3 (low) to 15 (high)) are used as covariates.
The data originate from multiple hospitals which can be categorised in three groups as: peripheral hospital without a neuro-surgical unit (\code{Status = 1}), peripheral hospital with a neuro-surgical unit (\code{Status = 2}), and academic medical centre (\code{Status = 3}).
}
\usage{data(trauma)}
\references{
Jonker M.A., Pazira H. and Coolen A.C.C. (2023). \emph{Bayesian Federated Inference for Statistical Models}, \emph{Statistics in Medicine}, Vol. 0(0), 0-0. <https://doi.org/10.48550/arXiv.2302.07677>

Draaisma J.M.Th, de Haan A.F.J., Goris R.J.A. (1989). \emph{Preventable Trauma Deaths in the Netherlands - A prospective Multicentre Study}, \emph{The journal of Trauma}, Vol. 29(11), 1552-1557.
}
\keyword{datasets}
